requirements.txt
asciitree==0.2
blosc==1.2.4
cffi==0.8.6
ipython==2.2.0
msgpack-python==0.4.2
py==1.4.23
pycparser==2.10
pyflakes==0.8.1
pytest==2.6.1
pyzmq==14.3.1
rethinkdb==1.14.0-1
six==1.7.3
-e git+git@github.com:deontologician/spaceship-build.git@bda337a1ba3865efc42b259826f11a4468194050#egg=spaceship_build-josh/clientserver
